Re-writing external layer in typescript.
